+++ NEWS.md
@@ -1,12 +1,13 @@
 ## 0.5.0 to 0.5.1 (unreleased so far)
 
- (TODO)
  * Contactdb: Adds search for annotations by tag-name.
  * Contactdb: Allows to configure the list of common "tags".
+ * Contactdb: Changes email search to be case-insensitive.
 
 ### Upgrade
  * Configuration: (optional) Add "common\_tags" to the contactdb config file,
      otherwise the default ones may be exposed to users by the frontend.
+
 
 ## 0.4.3.dev0 to 0.5.0
  * Contactdb: Allows tracking of db changes by user.